[flink] Add memory pool and allocator in PrepareCommitOperator
